+++ packages/supabase/lib/src/constants.dart
@@ -1,0 +1,2 @@
+const String integrationName = 'sentrySupabaseTracing';
+const String loggerName = 'sentry_supabase';